EO-894 - Rename alerts action/reducer/selector
