+++ src/reducers/index.js
@@ -6,7 +6,7 @@ import accessControlReady from './accessControlReady';
 import account from './account';
 import accountSingleSignOn from './accountSingleSignOn';
 import abTesting from './abTesting';
-import alertsV1 from './alertsV1';
+import alerts from './alerts';
 import apiKeys from './api-keys';
 import auth from './auth';
 import billing from './billing';
@@ -52,7 +52,7 @@ const appReducer = combineReducers({
   accessControlReady,
   account,
   accountSingleSignOn,
-  alertsV1,
+  alerts,
   auth,
   billing,
   bounceReport,
